Updated WaveConfig, created new path and wave, change in enemypathing
Waveconfig was updated to include getter methods to fetch the different variables from different prefabs / scripts.

A new path and wave was generated for different groups of enemies (to move on different paths)

We set the enemypathing script to fetch the waypoints from the prefab (rather than drag and drop) and set it just to the second wave for now.