Updated Enemy Spawner + Created DamageDealer and Enemy scripts
Enemy Spawner was updated to spawn all groups of enemies rather than just one. An option was also included to keep looping all of the groups so that the enemies can keep reappearing during the game.

DamageDealer: This script was created to set the laser damage settings.

Enemy: To handle the enemy's collisions with lasers and its health.