Updated Player and Enemy + Created Background Scroller and Particle system
Player and Enemy were updated so that the laser is destroyed once it hits an object and hits shouldn't be processed if they do not contain a DamageDealer component.

A new background was created and functionality was included to make it scroll.

A particle system was also created to create a star field visual effect.